third_party/WebKit/LayoutTests/fast/dom/shadow/get-destination-insertion-points-re-distribution-expected.txt

Tests for getDestinationInsertionPoints() which involves re-distribution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.item(1) is expectedNodes[1]
PASS successfullyParsed is true

TEST COMPLETE
